# openapi.yaml
openapi: 3.1.0
info:
  title: AAW-Skills
  description: AAW-Skills
  version: 1.0.0
servers:
  - url: '{scheme}://{host}:{port}/api/v1'
    variables:
      scheme:
        default: "http"
      hostname:
        default: "aaw.ngimenez.fr"
      port:
        default: "8000"
paths:
  "/auth":
    get:
      summary: Start the authorization process with Discord OAuth2
      description: HTTP redirect to the Discord OAuth2 authorization server
  "/auth/callback":
    get:
      summary: Callback which receives the authorization code from Discord
      description:
        Discord's authorization server redirects the user here. The query string contains the authorization code.
        The user is redirected to the front-end application.
        If the authorization successes, a cookie containing a JSON Web Token is written.
        If the authorization fails, `error` and `error_description` query string parameters are appended in the redirect URI.
      parameters:
        - name: code
          in: query
          description: The authorization code which is required to get a OAuth token
          required: true
          schema:
            type: string

  "/skills":
    summary: List of skills
    description: Get the list of skills
    get:
      summary: Get the list of skills
      responses:
        200:
          description: List of existing skills
          content:
            "application/json":
              schema:
                type: array
                items:
                  type: string

  "/skills/my":
    summary: User's skills
    description: Get or manage authenticated user's skills
    get:
      summary: Get authenticated user's skills
      security:
        - user: [ ]
      responses:
        200:
          description: List of skills with associated grade
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user_skill"
    put:
      summary: Set a skill for authenticated user
      security:
        - user: [ ]
      requestBody:
        content:
          "application/json":
            schema:
              $ref: '#/components/schemas/user_skill'
      responses:
        200:
          description: Update successful.
        400:
          description: Invalid request body.
        404:
          description: Unknown skill.
    delete:
      summary: Remove a skill for authenticated user
      security:
        - user: [ ]
      requestBody:
        content:
          "text/plain":
            description: Name of the skill to remove
      responses:
        200:
          description: Removed skill successfully.
        404:
          description: Unknown skill.

  /users:
    get:
      summary: List users
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: "#/components/parameters/limit"
        - name: skills
          in: query
          description: If true, the response will contain the skills property.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: List of users
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user_info"

  "/users/search/{query}":
    get:
      summary: Find users by name
      parameters:
      - name: query
        in: path
        description: First chars of the user name to look for.
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/offset'
      - $ref: "#/components/parameters/limit"
      responses:
        200:
          description: Results - list of users
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/basic_user_info"

  "/users/me":
    get:
      summary: Get information about authenticated user
      security:
        - user: [ ]
      responses:
        200:
          description: Information about authenticated user
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/user_info"
        401:
          description: Not authenticated

  "/users/{discord_id}":
    summary: Get information about a user
    parameters:
      - name: discord_id
        in: path
        schema:
          type: string
        required: true
    get:
      summary: Get information about a user
      responses:
        200:
          description: User information

  "/users/{discord_id}/skills":
    put:
      summary: Set a skill for given user
      description: (this endpoint can be used by the bot only)
      security:
        - bot: [ ]
      parameters:
        - name: discord_id
          in: path
          schema:
            type: string
          required: true
      requestBody:
        content:
          "application/json":
            schema:
              $ref: '#/components/schemas/user_skill'
      responses:
        200:
          description: Update successful.
        400:
          description: Invalid request body.
        404:
          description: Unknown skill or user.



components:
  parameters:
    offset:
      name: offset
      in: query
      description: Number of skipped results. Used for pagination.
      required: false
      schema:
        type: integer
        default: 0
    limit:
      name: count
      in: query
      description: Maximum number of results. Used for pagination.
      required: false
      schema:
        type: integer
        default: 25
  schemas:
    user_skill:
      type: object
      properties:
        name:
          description: Skill name
          type: string
        grade:
          description: Grade from 1 to 10 (included)
          type: integer
          exclusiveMinimum: 0
          maximum: 10
    basic_user_info:
      type: object
      properties:
        name:
          description: User name
          type: string
        discord_id:
          description: Discord ID (numeric string)
          type: string
        last_update:
          description: Last update date in French (France) format
          type: string
    user_info:
      type: object
      properties:
        name:
          description: User name
          type: string
        discord_id:
          description: Discord ID (numeric string)
          type: string
        last_update:
          description: Last update date in French (France) format
          type: string
        skills:
          description: List of skills
          type: array
          items:
            $ref: "#/components/schemas/user_skill"
  securitySchemes:
    user:
      type: http
      scheme: bearer
      description: JSON Web Token obtained using `/auth` endpoint
    bot:
      type: http
      scheme: basic
      description: |
        `bot:<password>`, encoded as Base64, where `<password>` value is defined in backend `.env` file